Fragmentation support for HCI/ACL packets
From Bluetooth Core Specs:

The LE_Read_Buffer_Size command is used to read the maximum size of the
data portion of HCI LE ACL Data Packets sent from the Host to the Controller.
The Host will segment the data transmitted to the Controller according to these
values, so that the HCI Data Packets will contain data with up to this size.

If the Controller returns a length value of zero, the Host shall use the
Read_Buffer_Size command to determine the size of the data buffers (shared
between BR/EDR and LE transports).